azure_deepresearch_agent/requirements.txt
chainlit==2.6.8
python-dotenv==1.1.1
azure-identity==1.24.0
azure-ai-projects==1.1.0b2
azure-ai-agents==1.2.0b1
azure-core==1.35.0

